{{cookiecutter.app_directory}}/requirements.txt
flask==3.1.0
gunicorn==23.0.0
jinja-partials==0.2.1
python-dotenv==1.0.1
